Update fun_game.py added battle code
4 total functions for handling battling
pokeChoose() allows you to choose a pokemon and returns either that pokemons object or a list of all of its attributes.

pokeStats() takes in a list of pokemon attributes and prints it to the terminal.

apManager() is for quickly adding or subtracting ap bonuses 

battleScenario() is pretty much the entire combat loop. It alternates attackers and defender by having whoevers turn it is be input as player1 who is always the attacker and whoevers turn it isnt as player2 who is always the defender. While the combat status is set to true the regular gameplay loop will repeat the battle part of the code so turns are still counting but players cant move away or catch new pokemon.

To the extent of debugging I've done it should work completely but could use some polish in some areas.
Also currently when a player wins it just prints out that that player won and then continues the game as before. I tested it only using the preset save.